Refactored code into src
- Updated Makefile
- Added git files
- Stuctured project close to a normal project
- Added a README
