+++ README.md
@@ -1,0 +1,7 @@
+### C++ AVL Tree
+
+While studying computer science at Dakota State University in 2021, I created this AVL Tree as part
+of an assignment for Advanced Data Structures.
+
+Original code was written during that class. After graduating in 2022, I refactored the project and
+updated the git structure. Feel free to use any of this code in learning AVL Trees in C++.
