+++ data/README.md
@@ -1,0 +1,10 @@
+# Data files
+
+`script.py` is a python script to generate a data/command file for the AVL Tree data input structure.
+
+### Variables
+`N_SIZE` = Total number of numbers to enter.  
+`MIN` = Minimum number (Inclusive).  
+`MAX` = Maximum number (Inclusive).  
+`FILE_NAME` = Name of file to output to. This will be entered as a command line argument.  
+`N_TO_DELETE` = Total number of numbers to delete. Chosen randomly.
