+++ src/customErrorClass.cpp
