+++ src/include/customErrorClass.h
